requirements.txt
Django==4.1.7
djangorestframework==3.14.0
django-filter==22.1
requests==2.28.2
djangorestframework-camel-case==1.4.2